requirements.txt
# Digital Kant Baseline v2.0 - 生产环境依赖
# Production Dependencies

# ============================================
# 核心框架
# ============================================
streamlit>=1.28.0
python-dotenv>=1.0.0

# ============================================
# LLM & AI 模型
# ============================================
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
torch>=2.0.0  # sentence-transformers 依赖
langchain-huggingface>=0.0.1  # HuggingFace embeddings for LangChain
accelerate>=0.25.0  # GPU acceleration utilities

# ============================================
# 向量数据库
# ============================================
chromadb>=0.4.15
pysqlite3-binary>=0.5.0  # SQLite 3.35+ 支持（ChromaDB 要求）

# ============================================
# 数据抓取与处理
# ============================================
beautifulsoup4>=4.12.0
lxml>=4.9.0  # BeautifulSoup 解析器
requests>=2.31.0
html5lib>=1.1  # 备用 HTML 解析器

# ============================================
# 文本处理
# ============================================
langdetect>=1.0.9  # 语言检测
regex>=2023.10.3  # 高级正则表达式
langchain-text-splitters>=0.0.1  # 文本切分（递归字符切分器）
tiktoken>=0.5.0  # Token 计数（用于精确的 Token 级别切分）

# ============================================
# 配置管理
# ============================================
pyyaml>=6.0.1

# ============================================
# 日志与监控
# ============================================
loguru>=0.7.2  # 现代化日志库

# ============================================
# 工具库
# ============================================
tqdm>=4.66.0  # 进度条
pydantic>=2.5.0  # 数据验证
pydantic-settings>=2.1.0  # 配置验证
